add commonForEachStatement
